+++ src/Codeception/Module/WebDriver.php
@@ -844,19 +844,26 @@ class WebDriver extends CodeceptionModule implements
         if (empty($nodes)) {
             $this->fail("No links containing text '$text' were found in page " . $this->_getCurrentUri());
         }
-        if (!$url) {
-            $this->assertNodesContain($text, $nodes, 'a');
-            return;
+        if ($url) {
+            $nodes = array_filter(
+                $nodes,
+                function (WebDriverElement $e) use ($url) {
+                    return trim($e->getAttribute('href')) == trim($url);
+                }
+            );
+            if (empty($nodes)) {
+                $this->fail("No links containing text '$text' and URL '$url' were found in page " . $this->_getCurrentUri());
+            }
         }
-        $this->assertNodesContain($text, $nodes, "a[href=$url]");
     }
-
 
     public function dontSeeLink($text, $url = null)
     {
         $nodes = $this->webDriver->findElements(WebDriverBy::partialLinkText($text));
         if (!$url) {
-            $this->assertNodesNotContain($text, $nodes, 'a');
+            if (!empty($nodes)) {
+                $this->fail("Link containing text '$text' was found in page " . $this->_getCurrentUri());
+            }
             return;
         }
         $nodes = array_filter(
@@ -865,7 +872,9 @@ class WebDriver extends CodeceptionModule implements
                 return trim($e->getAttribute('href')) == trim($url);
             }
         );
-        $this->assertNodesNotContain($text, $nodes, "a[href=$url]");
+        if (!empty($nodes)) {
+            $this->fail("Link containing text '$text' and URL '$url' was found in page " . $this->_getCurrentUri());
+        }
     }
 
     public function seeInCurrentUrl($uri)